separate code-editor view component
