+++ components/code-editor/index.ts
@@ -1,0 +1,2 @@
+export * from "./CodeEditor";
+export * from "./CodeEditorView";
